jimt/dance-ai-modal-refactoring - renamed ai/types -> ai/ai-types
